split phonon
